Rapidshare Hoster updated
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@2521 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/controlling/DistributeData.java
@@ -227,7 +227,7 @@ public class DistributeData extends ControlBroadcaster {
                             dLinks.get(c).setBrowserUrl(decrypted.getBrowserUrl());
                             dLinks.get(c).setProperties(decrypted.getProperties());
                             dLinks.get(c).getLinkStatus().setStatusText(decrypted.getLinkStatus().getStatusString());
-                            dLinks.get(c).setDownloadSize((int) decrypted.getDownloadSize());
+                            dLinks.get(c).setDownloadSize(decrypted.getDownloadSize());
 
                         }
